FDS/requirements.txt
#django and the third party
django==4.1
django-tinymce==3.5.0
django-crispy-forms==1.14.0
crispy_bootstrap5==0.7
django-environ==0.9.0
djangorestframework==3.14.0
django-filter==22.1
graphene-django==3.0.0
django-cors-headers==3.13.0

# uwsgi
# uwsgi==2.0.21

# mysql
# mysqlclient==2.2.0

# redis
django-redis==5.3.0
redis==4.6.0

# database
psycopg2

# display
fontawesomefree==6.2.1
Markdown==3.4.1

# image
pillow==9.4.0